# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: كُتَّاب الرحمن - نظام إدارة دار تحفيظ القرآن الكريم
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /circles:
    get:
      operationId: listCircles
      tags: [circles]
      summary: List all circles
      responses:
        "200":
          description: List of circles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Circle"
    post:
      operationId: createCircle
      tags: [circles]
      summary: Create a new circle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CircleInput"
      responses:
        "201":
          description: Created circle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circle"
  /circles/{id}:
    get:
      operationId: getCircle
      tags: [circles]
      summary: Get a single circle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Circle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circle"
    patch:
      operationId: updateCircle
      tags: [circles]
      summary: Update circle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CircleUpdate"
      responses:
        "200":
          description: Updated circle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circle"
    delete:
      operationId: deleteCircle
      tags: [circles]
      summary: Delete a circle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── MEMORIZATION ────────────────────────────────────────────
  /memorization:
    get:
      operationId: listMemorization
      tags: [memorization]
      summary: List memorization records (filter by studentId or date)
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of memorization records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemorizationRecord"
    post:
      operationId: createMemorization
      tags: [memorization]
      summary: Record daily memorization for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorizationInput"
      responses:
        "201":
          description: Created memorization record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemorizationRecord"
  /memorization/{id}:
    patch:
      operationId: updateMemorization
      tags: [memorization]
      summary: Update a memorization record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorizationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemorizationRecord"
    delete:
      operationId: deleteMemorization
      tags: [memorization]
      summary: Delete a memorization record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ATTENDANCE ──────────────────────────────────────────────
...
  /guardians:
    get:
      operationId: listGuardians
      tags: [guardians]
      summary: List guardians
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Guardians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guardian"
    post:
      operationId: createGuardian
      tags: [guardians]
      summary: Add guardian info for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardianInput"
      responses:
        "201":
          description: Guardian created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guardian"
  /guardians/{id}:
    patch:
      operationId: updateGuardian
      tags: [guardians]
      summary: Update guardian info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardianUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guardian"

  # ─── PAYMENTS ────────────────────────────────────────────────
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List kuttab expenses
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Record an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Expense recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CERTIFICATES ────────────────────────────────────────────
...
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List exams and competitions
      responses:
        "200":
          description: Exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create an exam or competition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "201":
          description: Exam created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{id}:
    delete:
      operationId: deleteExam
      tags: [exams]
      summary: Delete an exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /exam-results:
    get:
      operationId: listExamResults
      tags: [exams]
      summary: List exam results
      parameters:
        - name: examId
          in: query
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Exam results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamResult"
    post:
      operationId: createExamResult
      tags: [exams]
      summary: Record an exam result for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamResultInput"
      responses:
        "201":
          description: Result recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"

  # ─── ANALYTICS ───────────────────────────────────────────────
...
components:
  schemas:
...
    Circle:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        schedule:
          type: ["string", "null"]
        onlineLink:
          type: ["string", "null"]
        studentCount:
          type: integer
        createdAt:
          type: string
    CircleInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        schedule:
          type: string
        onlineLink:
          type: string
    CircleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        schedule:
          type: string
        onlineLink:
          type: string

    # ─── MEMORIZATION SCHEMAS ─────────────────────────────────
    MemorizationRecord:
      type: object
      required: [id, studentId, studentName, surahName, fromVerse, toVerse, rating, date, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        surahName:
          type: string
        fromVerse:
          type: integer
        toVerse:
          type: integer
        rating:
          type: string
          enum: [ممتاز, جيد جداً, جيد, مقبول, ضعيف]
        notes:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
    MemorizationInput:
      type: object
      required: [studentId, surahName, fromVerse, toVerse, rating, date]
      properties:
        studentId:
          type: integer
        surahName:
          type: string
        fromVerse:
          type: integer
        toVerse:
          type: integer
        rating:
          type: string
        notes:
          type: string
        date:
          type: string
    MemorizationUpdate:
      type: object
      properties:
        surahName:
          type: string
        fromVerse:
          type: integer
        toVerse:
          type: integer
        rating:
          type: string
        notes:
          type: string

    # ─── ATTENDANCE SCHEMAS ───────────────────────────────────
...
    Guardian:
      type: object
      required: [id, studentId, studentName, name, relationship, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        name:
          type: string
        relationship:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    GuardianInput:
      type: object
      required: [studentId, name, relationship]
      properties:
        studentId:
          type: integer
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
        email:
          type: string
        notes:
          type: string
    GuardianUpdate:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
        email:
          type: string
        notes:
          type: string

    # ─── PAYMENT SCHEMAS ──────────────────────────────────────
...
    Expense:
      type: object
      required: [id, title, amount, category, date, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [title, amount, category, date]
      properties:
        title:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string
        notes:
          type: string

    # ─── CERTIFICATE SCHEMAS ──────────────────────────────────
...
    Exam:
      type: object
      required: [id, title, type, date, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [اختبار, مسابقة]
        date:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    ExamInput:
      type: object
      required: [title, type, date]
      properties:
        title:
          type: string
        type:
          type: string
        date:
          type: string
        description:
          type: string
    ExamResult:
      type: object
      required: [id, examId, examTitle, studentId, studentName, score, createdAt]
      properties:
        id:
          type: integer
        examId:
          type: integer
        examTitle:
          type: string
        studentId:
          type: integer
        studentName:
          type: string
        score:
          type: number
        grade:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ExamResultInput:
      type: object
      required: [examId, studentId, score]
      properties:
        examId:
          type: integer
        studentId:
          type: integer
        score:
          type: number
        grade:
          type: string
        notes:
          type: string

    # ─── ANALYTICS SCHEMAS ────────────────────────────────────
...