# lib/api-zod/src/generated/types/circle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * كُتَّاب الرحمن - نظام إدارة دار تحفيظ القرآن الكريم
 * OpenAPI spec version: 0.1.0
 */

export interface Circle {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  schedule?: string | null;
  /** @nullable */
  onlineLink?: string | null;
  studentCount?: number;
  createdAt: string;
}