# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * كُتَّاب الرحمن - نظام إدارة دار تحفيظ القرآن الكريم
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  title: string;
  amount: number;
  category: string;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}