# lib/api-zod/src/generated/types/guardian.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * كُتَّاب الرحمن - نظام إدارة دار تحفيظ القرآن الكريم
 * OpenAPI spec version: 0.1.0
 */

export interface Guardian {
  id: number;
  studentId: number;
  studentName: string;
  name: string;
  relationship: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}